+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file is a history of the changes made to idearium-lib.
 
+## Unreleased
+
+- Added `Logger` instance to the common folder.
+- Added new `util` to the library.
+- Added new `mongodb` to the common folder.
+- Added new `publisher` to the common folder.
+- Added new `opbeat` to the common folder.
+- Added new `session` to the common folder.
+- Deprecated `utils` in favour of the new `util`. This also acts as a wrapper around the build in node `util` module.
+
 ## 1.0.0-alpha.11
 
 - Large refactoring of the `mq.Connection` class.